Add Preditor, paramiter refinements
preditor added, travels to the birds percieved center. balance was had
to be found between the birds reaction radius and the amount of reaction
they took.

preditor moving quite slowly, would be better to have the preditor
move to a point beyond the starlings so that it doesn't slow down as
much and starlings move away better.